requirements.txt
opencv-python
numpy
stegano
deepface
sqlite3
hashlib
flask
requests
pillow
